Move package to the block folder
- Move whole `packages/tiled-grid` to Tiled gallery block folder
- Remove Prettier pragma
- Remove debug code
