+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/constants.js
@@ -1,4 +1,2 @@
-/** @format */
-
 export const CONTENT_WIDTH = 640; // same as example.png
 export const TILE_MARGIN = 4;
