+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/grouper.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * External dependencies
  */
