+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/item.js
@@ -1,5 +1,3 @@
-/** @format */
-
 export class Jetpack_Tiled_Gallery_Item {
 	// eslint-disable-next-line no-unused-vars
 	constructor( attachment_image, needs_attachment_link, grayscale ) {
