+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/rectangular-item.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * Internal dependencies
  */
