+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/row.js
@@ -1,5 +1,3 @@
-/** @format */
-
 export class Jetpack_Tiled_Gallery_Row {
 	constructor( groups ) {
 		this.groups = groups;
