+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/shapes/five.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * Internal dependencies
  */
