+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/shapes/jetpack-tiled-gallery-shape.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * External dependencies
  */
