+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/shapes/long-symmetric-row.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * Internal dependencies
  */
