+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/shapes/test/jetpack-tiled-gallery-shape.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * Internal dependencies
  */
