+++ client/gutenberg/extensions/tiled-gallery/tiled-grid/shapes/two.js
@@ -1,5 +1,3 @@
-/** @format */
-
 /**
  * Internal dependencies
  */
