--- packages/tiled-grid/.eslintrc.js
@@ -1,7 +1,0 @@
-/** @format */
-
-module.exports = {
-	rules: {
-		'import/no-extraneous-dependencies': [ 'error', { packageDir: __dirname } ],
-	},
-};
