--- packages/tiled-grid/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@automattic/tiled-grid",
-  "private": "true",
-  "dependencies": {
-	  "lodash": "^4.17.11"
-  },
-  "devDependencies": {
-    "deep-freeze": "^0.0.1",
-    "@babel/runtime": "^7.2.0",
-    "concurrently": "^4.1.0",
-    "nodemon": "^1.18.8"
-  },
-  "scripts": {
-    "build": "../../bin/sdk-cli.js generic index.js dist/index.js --watch",
-    "output": "nodemon dist/index.js",
-    "start": "concurrently -n build,output npm:build npm:output"
-  }
-}
